actually, killing off flag flipping entirely
