--- tools/deprecate_protos/deprecate_protos.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-. tools/shell_utils.sh
-
-set -e
-
-python_venv deprecate_protos
