--- tools/deprecate_protos/requirements.txt
@@ -1,1 +1,0 @@
-six==1.12.0
